# chore(parser): add benchmark with real-world doc