add the `verts-examples.adoc` file from
https://github.com/vert-x3/vertx-examples/blob/master/core-examples/README.adoc
and run the parser 50x

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>