@xcoulon 2019-07-02
Thanks for the feedback @moorereason, this looks actually great! 🙌 

Would you like to open a PR with your refactoring, maybe? You could even reference the commands in the [README.adoc](https://github.com/bytesparadise/libasciidoc/blob/master/README.adoc) or even better in the [CONTRIBUTE.adoc](https://github.com/bytesparadise/libasciidoc/blob/master/CONTRIBUTE.adoc) so we can easily measure future changes/refactorings. WDYT?